Akidev

+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/manga/MangaTrack.kt
@@ -8,7 +8,7 @@ interface MangaTrack : Serializable {
 
     var manga_id: Long
 
-    var tracker_id: Int
+    var tracker_id: Long
 
     var remote_id: Long
 
@@ -16,13 +16,13 @@ interface MangaTrack : Serializable {
 
     var title: String
 
-    var last_chapter_read: Float
+    var last_chapter_read: Double
 
-    var total_chapters: Int
+    var total_chapters: Long
 
-    var score: Float
+    var score: Double
 
-    var status: Int
+    var status: Long
 
     var started_reading_date: Long
 
@@ -40,7 +40,7 @@ interface MangaTrack : Serializable {
 
     companion object {
         fun create(serviceId: Long): MangaTrack = MangaTrackImpl().apply {
-            tracker_id = serviceId.toInt()
+            tracker_id = serviceId
         }
     }
 }
